requirements.txt
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
css-html-js-minify==2.5.5
cssselect==1.2.0
cssutils==2.11.1
dnspython==2.7.0
Faker==33.1.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-PyMongo==2.3.0
htmlmin==0.1.12
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
Levenshtein==0.26.1
lxml==5.3.0
MarkupSafe==3.0.2
more-itertools==10.5.0
numpy==2.0.2
pandas==2.2.3
pillow==11.0.0
premailer==3.10.0
PyJWT==2.10.1
pymongo==4.10.1
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-Levenshtein==0.26.1
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
RapidFuzz==3.11.0
reportlab==4.2.5
reportlib==3.4.0
requests==2.32.3
six==1.17.0
treepoem==3.25.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
zipp==3.21.0
